+++ prisma/seed/arts.ts
@@ -1,8 +1,8 @@
 import { Arts, PrismaClient } from '@prisma/client';
 const prisma = new PrismaClient();
 
-import { ART_COMMON_DATA, ART_IDS } from './data/art';
-import { AUTHOR_IDS } from './data/author';
+import { ART_COMMON_DATA, ART_IDS } from './data/arts';
+import { AUTHOR_IDS } from './data/authors';
 
 export const art = async () => {
 	const artsData = ART_IDS.map((art_id, index) => ({
