+++ prisma/seed/authors.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client';
-import { AUTHOR_COMMON_DATA, AUTHOR_IDS } from './data/author';
+import { AUTHOR_COMMON_DATA, AUTHOR_IDS } from './data/authors';
 
 const prisma = new PrismaClient();
