+++ prisma/seed/data/arts.ts
